requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
numpy
pgvector
python-dotenv
sqlalchemy
asyncpg
transformers
psycopg2-binary
requests
newspaper3k
lxml[html_clean]
SQLAlchemy